Trying to come up with a generic CRUD trait

+++ src/main/scala/coffeecraft/models/package.scala
@@ -1,5 +1,5 @@
 package coffeecraft
 
 package object models {
-  type CoffeeKey = Int
+  type CoffeeKey = Long
 }
